Make invalid items unusable (#7506)
