+++ test/player_test.h
@@ -10,9 +10,9 @@
 
 using namespace devilution;
 
-static int CountItems(Item *items, int n)
+static int CountItems(devilution::Item *items, int n)
 {
-	return std::count_if(items, items + n, [](Item x) { return !x.isEmpty(); });
+	return std::count_if(items, items + n, [](devilution::Item x) { return !x.isEmpty(); });
 }
 
 static int Count8(int8_t *ints, int n)
